Move tests of --last tag to basicUsageTests
